TODO finish copyTo in MapGroup1 and replace MapGroup1.encodeP in line 482 of SnoveEngine.

+++ core/src/main/java/org/bouncycastle/pqc/crypto/snova/GF16Utils.java
@@ -117,7 +117,7 @@ public class GF16Utils
         // If there is an extra nibble (odd number of nibbles), store it directly in lower 4 bits.
         if ((mlen & 1) == 1)
         {
-            menc[i] = (byte)m[i];
+            menc[i] = m[i];
         }
     }
